Emotion optimizations.
